Refactor release path discovery into shared helper
Extract FindReleasePath() and getUserHomeDir() into release_path.go
so that debug_ctr.go and server.go share the same logic for locating
the miren release directory.
